rename Warpipe -> FirePlant and add ani FirePlant

+++ 05-SceneManager/Pipe.h
@@ -23,7 +23,7 @@ public:
 	}
 	void RenderBoundingBox();
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
-	virtual int GetWidth() { return width; };
+	virtual float GetWidth() { return width; };
 	virtual void Render();
 	virtual int IsCollidable() { return 1; };
 	//virtual int IsBlockingOnSide(){ return 0; };
